kernel: move write to binary to capsules
