+++ capsules/system/src/lib.rs
@@ -13,3 +13,4 @@ pub mod scheduler;
 pub mod storage_permissions;
 pub mod syscall_filter;
 pub mod virtual_scheduler_timer;
+pub mod write_to_binary;
